# artifacts/api-server/src/routes/stats.ts
import { Router, type IRouter } from "express";
import { getDB } from "../lib/mongodb";

const router: IRouter = Router();

router.get("/stats/overview", async (_req, res): Promise<void> => {
  const db = getDB();
  const [totalMahasiswa, totalDosen, totalMataKuliah, mahasiswaAktif, dosenAktif] = await Promise.all([
    db.collection("mahasiswa").countDocuments(),
    db.collection("dosen").countDocuments(),
    db.collection("matakuLiah").countDocuments(),
    db.collection("mahasiswa").countDocuments({ status: "aktif" }),
    db.collection("dosen").countDocuments({ status: "aktif" }),
  ]);
  const prodi = await db.collection("mahasiswa").distinct("prodi");
  const nilaiDocs = await db.collection("nilai").find({}).toArray();
  const rataIpk = nilaiDocs.length > 0
    ? nilaiDocs.reduce((sum, n) => sum + ((n["nilaiAkhir"] as number) || 0), 0) / nilaiDocs.length / 25
    : 3.5;
  res.json({
    totalMahasiswa,
    totalDosen,
    totalMataKuliah,
    totalProdi: prodi.length,
    mahasiswaAktif,
    dosenAktif,
    rataIpk: Math.min(4.0, parseFloat(rataIpk.toFixed(2))),
    tingkatKelulusan: 87.5,
    mahasiswaBaru: Math.floor(totalMahasiswa * 0.15),
    dosenBaru: 2,
  });
});

router.get("/stats/mahasiswa", async (_req, res): Promise<void> => {
  const db = getDB();
  const [byProdiRaw, byStatusRaw, allMahasiswa] = await Promise.all([
    db.collection("mahasiswa").aggregate([{ $group: { _id: "$prodi", total: { $sum: 1 } } }]).toArray(),
    db.collection("mahasiswa").aggregate([{ $group: { _id: "$status", total: { $sum: 1 } } }]).toArray(),
    db.collection("mahasiswa").find({}).toArray(),
  ]);
  const byAngkatanMap: Record<number, number> = {};
  for (const m of allMahasiswa) {
    const a = m["angkatan"] as number;
    byAngkatanMap[a] = (byAngkatanMap[a] || 0) + 1;
  }
  res.json({
    byProdi: byProdiRaw.map((r) => ({ prodi: r["_id"] as string, total: r["total"] as number })),
    byStatus: byStatusRaw.map((r) => ({ status: r["_id"] as string, total: r["total"] as number })),
    byAngkatan: Object.entries(byAngkatanMap).map(([angkatan, total]) => ({ angkatan: parseInt(angkatan, 10), total })).sort((a, b) => a.angkatan - b.angkatan),
  });
});

router.get("/stats/akademik", async (_req, res): Promise<void> => {
  const db = getDB();
  const [allNilai, byProdiRaw] = await Promise.all([
    db.collection("nilai").find({}).toArray(),
    db.collection("mahasiswa").aggregate([{ $group: { _id: "$prodi", ipkSum: { $sum: "$ipk" }, count: { $sum: 1 } } }]).toArray(),
  ]);
  const gradeMap: Record<string, number> = {};
  for (const n of allNilai) {
    const g = (n["grade"] as string) || "E";
    gradeMap[g] = (gradeMap[g] || 0) + 1;
  }
  const total = allNilai.length || 1;
  const distribusiNilai = Object.entries(gradeMap).map(([grade, count]) => ({
    grade,
    total: count,
    persentase: parseFloat(((count / total) * 100).toFixed(1)),
  })).sort((a, b) => a.grade.localeCompare(b.grade));

  const rataIpkByProdi = byProdiRaw.map((r) => ({
    prodi: r["_id"] as string,
    rataIpk: parseFloat(((r["ipkSum"] as number) / ((r["count"] as number) || 1)).toFixed(2)),
  }));

  const absensiTotal = await db.collection("absensi").countDocuments();
  const absensiHadir = await db.collection("absensi").countDocuments({ status: "hadir" });
  const tingkatAbsensi = absensiTotal > 0 ? parseFloat(((absensiHadir / absensiTotal) * 100).toFixed(1)) : 92.5;

  res.json({ distribusiNilai, rataIpkByProdi, tingkatAbsensi });
});

router.get("/stats/absensi-summary", async (req, res): Promise<void> => {
  const db = getDB();
  const { userId, semester } = req.query as Record<string, string>;
  const filter: Record<string, unknown> = {};
  if (userId) filter["mahasiswaId"] = userId;

  const allAbsensi = await db.collection("absensi").find(filter).toArray();
  const totalPertemuan = allAbsensi.length;
  const hadir = allAbsensi.filter((a) => a["status"] === "hadir").length;
  const izin = allAbsensi.filter((a) => a["status"] === "izin").length;
  const sakit = allAbsensi.filter((a) => a["status"] === "sakit").length;
  const alpha = allAbsensi.filter((a) => a["status"] === "alpha").length;
  const persentaseKehadiran = totalPertemuan > 0 ? parseFloat(((hadir / totalPertemuan) * 100).toFixed(1)) : 100;

  const mkMap: Record<string, { hadir: number; total: number }> = {};
  for (const a of allAbsensi) {
    const mk = (a["mataKuliahNama"] as string) || "Unknown";
    if (!mkMap[mk]) mkMap[mk] = { hadir: 0, total: 0 };
    mkMap[mk].total++;
    if (a["status"] === "hadir") mkMap[mk].hadir++;
  }
  const byMataKuliah = Object.entries(mkMap).map(([mataKuliah, val]) => ({
    mataKuliah,
    hadir: val.hadir,
    total: val.total,
    persentase: parseFloat(((val.hadir / val.total) * 100).toFixed(1)),
  }));

  res.json({ totalPertemuan, hadir, izin, sakit, alpha, persentaseKehadiran, byMataKuliah });
});

router.get("/stats/ipk-trend", async (req, res): Promise<void> => {
  const db = getDB();
  const { mahasiswaId } = req.query as Record<string, string>;
  const filter: Record<string, unknown> = {};
  if (mahasiswaId) filter["mahasiswaId"] = mahasiswaId;

  const allNilai = await db.collection("nilai").find(filter).toArray();
  const semMap: Record<string, { total: number; count: number; sks: number }> = {};
  for (const n of allNilai) {
    const sem = (n["semester"] as string) || "Unknown";
    const nilaiAkhir = (n["nilaiAkhir"] as number) || 0;
    const gradePoint = nilaiAkhir >= 90 ? 4.0 : nilaiAkhir >= 85 ? 3.7 : nilaiAkhir >= 80 ? 3.3 : nilaiAkhir >= 75 ? 3.0 : nilaiAkhir >= 70 ? 2.7 : nilaiAkhir >= 65 ? 2.3 : nilaiAkhir >= 60 ? 2.0 : nilaiAkhir >= 55 ? 1.0 : 0;
    if (!semMap[sem]) semMap[sem] = { total: 0, count: 0, sks: 0 };
    semMap[sem].total += gradePoint;
    semMap[sem].count++;
    semMap[sem].sks += (n["sks"] as number) || 0;
  }

  const data = Object.entries(semMap).map(([semester, val]) => ({
    semester,
    ipk: parseFloat((val.total / (val.count || 1)).toFixed(2)),
    sks: val.sks,
  }));

  const mahasiswa = mahasiswaId ? await db.collection("mahasiswa").findOne({ _id: undefined }) : null;
  res.json({ data, currentIpk: data.length > 0 ? data[data.length - 1].ipk : 3.5, totalSks: data.reduce((sum, d) => sum + d.sks, 0) });
});

router.get("/stats/jadwal-today", async (req, res): Promise<void> => {
  const db = getDB();
  const { userId } = req.query as Record<string, string>;
  const today = new Date();
  const hariList = ["Minggu", "Senin", "Selasa", "Rabu", "Kamis", "Jumat", "Sabtu"];
  const todayHari = hariList[today.getDay()];
  const currentTime = `${today.getHours().toString().padStart(2, "0")}:${today.getMinutes().toString().padStart(2, "0")}`;

  const filter: Record<string, unknown> = { hari: todayHari };
  if (userId) filter["dosenId"] = userId;

  const jadwal = await db.collection("jadwal").find(filter).toArray();
  const processed = jadwal.map((j) => ({ ...j, id: (j["_id"] as { toHexString(): string }).toHexString() }));

  const selesai = processed.filter((j) => (j["jamSelesai"] as string) < currentTime).length;
  const berlangsung = processed.filter((j) => (j["jamMulai"] as string) <= currentTime && (j["jamSelesai"] as string) >= currentTime).length;
  const belumMulai = processed.filter((j) => (j["jamMulai"] as string) > currentTime).length;

  res.json({
    totalJadwal: jadwal.length,
    selesai,
    berlangsung,
    belumMulai,
    jadwal: processed.map(({ _id, ...rest }) => rest),
  });
});

export default router;