# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Kampus Modern API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stats/absensi-summary:
    get:
      operationId: getStatsAbsensiSummary
      tags: [stats]
      summary: Get absensi summary for a user
      parameters:
        - name: userId
          in: query
          schema:
            type: string
        - name: semester
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Absensi summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsAbsensiSummary"
  /stats/ipk-trend:
    get:
      operationId: getStatsIpkTrend
      tags: [stats]
      summary: Get IPK trend over semesters
      parameters:
        - name: mahasiswaId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: IPK trend data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsIpkTrend"
  /stats/jadwal-today:
    get:
      operationId: getStatsJadwalToday
      tags: [stats]
      summary: Get today's schedule count and overview
      parameters:
        - name: userId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Today jadwal overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsJadwalToday"
components:
  schemas:
...
    Jadwal:
      type: object
      properties:
        id:
          type: string
        mataKuliahId:
          type: string
        mataKuliahNama:
          type: string
        mataKuliahKode:
          type: string
        dosenId:
          type: string
        dosenNama:
          type: string
        hari:
          type: string
          enum: [Senin, Selasa, Rabu, Kamis, Jumat, Sabtu]
        jamMulai:
          type: string
        jamSelesai:
          type: string
        ruangan:
          type: string
        kelas:
          type: string
        semester:
          type: string
        sks:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - mataKuliahId
        - dosenId
        - hari
        - jamMulai
        - jamSelesai
        - ruangan
        - kelas
        - semester
...
    StatsAbsensiSummary:
      type: object
      properties:
        totalPertemuan:
          type: integer
        hadir:
          type: integer
        izin:
          type: integer
        sakit:
          type: integer
        alpha:
          type: integer
        persentaseKehadiran:
          type: number
        byMataKuliah:
          type: array
          items:
            type: object
            properties:
              mataKuliah:
                type: string
              hadir:
                type: integer
              total:
                type: integer
              persentase:
                type: number
            required:
              - mataKuliah
              - hadir
              - total
              - persentase
      required:
        - totalPertemuan
        - hadir
        - izin
        - sakit
        - alpha
        - persentaseKehadiran
    StatsIpkTrend:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
            properties:
              semester:
                type: string
              ipk:
                type: number
              sks:
                type: integer
            required:
              - semester
              - ipk
              - sks
        currentIpk:
          type: number
        totalSks:
          type: integer
      required:
        - data
        - currentIpk
        - totalSks
    StatsJadwalToday:
      type: object
      properties:
        totalJadwal:
          type: integer
        selesai:
          type: integer
        berlangsung:
          type: integer
        belumMulai:
          type: integer
        jadwal:
          type: array
          items:
            $ref: "#/components/schemas/Jadwal"
      required:
        - totalJadwal
        - selesai
        - berlangsung
        - belumMulai
        - jadwal